Hircine.Command

Description

This module provides a typed interface for handling IRC commands. It lets you work in terms of e.g. Join or PrivMsg values instead of having to match on the raw Command type.

Parsing and un-parsing commands

class IsCommand a where

Represents a type that can be converted to and from a raw Command.

Minimal complete definition

fromCommand, toCommand

Methods

fromCommand :: Command -> Maybe a

toCommand :: a -> Command

type Bytes = ByteString

Convenient synonym for ByteString.

Command types

pattern Join :: [Bytes] -> Maybe [Bytes] -> ParsedCommand "JOIN" (CommaSep Bytes, Maybe (CommaSep Bytes))

pattern Mode :: Bytes -> [Bytes] -> ParsedCommand "MODE" (Bytes, [Bytes])

pattern Nick :: Bytes -> ParsedCommand "NICK" Bytes

pattern Notice :: [Bytes] -> Bytes -> ParsedCommand "NOTICE" (CommaSep Bytes, Bytes)

pattern Pass :: Bytes -> ParsedCommand "PASS" Bytes

pattern Ping :: Bytes -> Maybe Bytes -> ParsedCommand "PING" (Bytes, Maybe Bytes)

pattern Pong :: Bytes -> Maybe Bytes -> ParsedCommand "PONG" (Bytes, Maybe Bytes)

pattern PrivMsg :: [Bytes] -> Bytes -> ParsedCommand "PRIVMSG" (CommaSep Bytes, Bytes)

pattern Quit :: Maybe Bytes -> ParsedCommand "QUIT" (Maybe Bytes)

pattern User :: Bytes -> Bytes -> ParsedCommand "USER" (Bytes, Bytes, Bytes, Bytes)

Low-level machinery

This section contains the low-level machinery that makes this module work. You shouldn't need to deal with these routines unless you're adding a new command.

newtype ParsedCommand method params

Represents a validated IRC command.

Constructors

ParsedCommand params

type ParamParser = StateT [Bytes] Maybe

runParamParser :: Alternative f => ParamParser a -> [Bytes] -> f a

class IsParams a where

Minimal complete definition

parseParams, renderParams

Methods

parseParams :: ParamParser a

renderParams :: a -> [Bytes]

newtype CommaSep a

Constructors

CommaSep [a]

unCommaSep :: CommaSep a -> [a]