Hircine.Core

Types

data Message

An IRC message consists of an optional prefix, along with a message body.

Constructors

Message

Fields

msgOrigin :: !(Maybe Origin)

msgCommand :: !Command

data Origin

A message can originate from a fellow user, or the server itself.

If the former, the prefix must specify the nickname, username and host address. Technically only the nickname is required, but in practice servers always supply all three fields.

Constructors

FromServer !ByteString

FromUser !ByteString !ByteString !ByteString

data Command

An IRC command, along with its parameters.

Invariant: Only the final parameter may contain whitespace, or start with a colon (:).

Constructors

Command

Fields

cmdMethod :: !Method

cmdParams :: ![ByteString]

data Method

A command is specified by either a sequence of uppercase ASCII letters, or a three digit code.

Constructors

Textual !ByteString

Numeric !Word8 !Word8 !Word8

Parsing and rendering

parseMessage :: ByteString -> Either String Message

Parse a message. The input string should represent exactly one message, with no trailing newlines.

renderMessage :: Message -> ByteString

Render a message into a string.

renderCommand :: Command -> ByteString

showMessage :: Message -> String

showCommand :: Command -> String

Utilities

decodeIrc :: ByteString -> Text

Try to decode using UTF-8, then Latin-1.